Deduplicate tokenItems components into shared composables and tx helpers
Extracts the verbatim-duplicated code identified in code review issue 18:

- utils/txHelpers.ts: confirmDialog, notifySending, showTransactionResult
  (replaces 12 dialog promise wrappers and 8 tx-success tails)
- utils/tokenComposables.ts: useTokenAddressInput incl. validateDestination,
  and useNftParsing for parsable-NFT state
- tokenItems/nftMintForm.vue: mint form previously duplicated in
  tokenItemNFT.vue and nftChild.vue, sharing the host action lock
- utils.ts parseTokenAmountToBigInt: replaces the 3x pasted amount pipeline
  in tokenItemFT.vue using string math, fixing the float-precision finding
- local handleTransactionError copies replaced by existing displayAndLogError

Behavior changes: FT burn now updates wallet history like other tx flows,
FT auth-transfer validation errors now display as toasts instead of
unhandled rejections, and the mint form has its own destination field.
